Hide raptor's unsaved edit warning message.
It hangs around after saving the page because it's outside of the refreshed
area. There's no quick way to fix that currently and we in any case want to
move to pat-autosubmit for saving the form, in which case it won't be necessary
any more because we'll be using our own mechanism (similar to star).