Reword the workflow select
The current way in which pat-select and the workflow dropdown functioned is not
representative to how workflow changing happens in Plone.

In Plone, you don't have a list of state names to choose from, you have a list
of so-called "transitions" to choose from, and when you choose a transition,
you move from one state to another.

I'm not aware of an easy way to tell which state a transition will lead to.

I've therefore modified the proto a bit to show the fact that the wofkflow
dropdown shows a state in its label, but when you open the select you get a
list of transitions.