Icon for external editing #77
@bfackland , I’m not convinced is my solution will be very clear,
unless people read the tooltip. Would it be possible to a different
icon depending on the file type? So the Word icon for Word, the PP icon
for PP (I have all those in my set and the current icon as a fallback
for files we have no icon for.